Add auto_update for sts-silicom app.
This commit adds the auto update capability to the sts-silicom app.
This change will enable the sts-silicom app to be updated when a
new version is made available to the system.

PASS: build-pkgs -p stx-sts-silicom-helm
PASS: system application-apply sts-silicom
PASS: In a system where sts-silicom is in applied state, verify
      whether auto_update works by placing the new tarball
      in "/usr/local/share/applications/helm/" and monitoring
      if sysinv automaticaly updates sts-silicom to a new version.

Story: 2010628
Task: 48742

Change-Id: I055455dbd7f1bb0fd96d10eac774e57f4eeb7fda
Signed-off-by: David Barbosa Bastos <david.barbosabastos@windriver.com>